build: fix arm64 compilation and linking issues for Vulkan server
- Wrap x86_64-specific GSBASE CPU context patching in schedcoop with defined(__x86_64__)
- Implement conditional configuration space reading using pci_generic_config_read on arm64 in virtio_pci
- Export uk_venus_query_physical_device_properties and pci_generic_config_read in their respective exportsyms
- Fix static scoping, implicit declarations, and print format warnings in libvulkan runtime
- Update related unikraft patches (0001, 0004) to persist arm64 fixes
- Verified that both `make test-fast` and `make llama-vk-server-build` pass successfully for target qemu/arm64

+++ libs/libvulkan/runtime/vk_pipeline.c
@@ -21,7 +21,7 @@ VkResult stub_vkCreateShaderModule(VkDevice dev, const void *ci,
                                          (uint32_t)(code_size / 4u));
     printf("VOGUE-DBG createShaderModule h=0x%llx codeSize=%llu encpos=%u ovf=%d bufsz=%u\n",
            (unsigned long long)h, (unsigned long long)code_size,
-           _enc.pos, _enc.overflow, UK_DISPATCH_BUF_SIZE);
+           (unsigned int)_enc.pos, _enc.overflow, UK_DISPATCH_BUF_SIZE);
     UK_ENC_SUBMIT();
     *pShader = (VkShaderModule)h;
     return VK_SUCCESS;
